requirements.txt
discord.py[voice]==2.6.3
discord-ext-voice-recv>=0.1.7a94
SpeechRecognition==3.8.1
google-generativeai>=0.8.5
python-dotenv==1.0.0
Flask
